==> 3_interpolate_aq_data.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt

data = pd.read_csv("./all_aq_data.csv")
data.drop(columns='Unnamed: 0', inplace=True)

data.groupby(['Site.ID', 'school_year']).ngroups

# Convert 'Date' column to datetime format
data['Date'] = pd.to_datetime(data['Date'], format='%Y/%m/%d')

# Sort the DataFrame by 'Site.ID' and 'Date'
data.sort_values(['Site.ID', 'Date'], inplace=True)

valid_sites = data

valid_sites.groupby(['Site.ID', 'school_year']).ngroups

valid_sites[valid_sites['school_year'] == 2002]

def interpolate_monthly(group):
    # Ensure the group is sorted by Date
    group = group.sort_values('Date')

    # Get the year and month from the first row of the group
    year_month = group['Date'].dt.to_period('M').iloc[0]
    start_of_month = year_month.start_time
    end_of_month = year_month.end_time

    # Create a date range for the full month
    full_month_range = pd.date_range(start=start_of_month, end=end_of_month)

    # Set the index to the date for resampling purposes
    group = group.set_index('Date')

    # Reindex the group with the full month date range, introducing NaNs for missing dates
    group = group.reindex(full_month_range)

    # Fill the 'year' and 'month' columns
    group['year'] = group.index.year
    group['month'] = group.index.month

    # Backward fill the site information columns where we have NaNs due to reindexing
    group[['Site.ID', 'SITE_LATITUDE', 'SITE_LONGITUDE', 'school_year']] = group[['Site.ID', 'SITE_LATITUDE', 'SITE_LONGITUDE', 'school_year']].bfill().ffill()

    # Calculate the maximum gap of consecutive missing days
    # This will include gaps at the beginning or the end of the month
    group['consecutive_missing'] = group['Daily.Mean.PM2.5.Concentration'].isna().astype(int).groupby(group['Daily.Mean.PM2.5.Concentration'].notna().astype(int).cumsum()).cumsum()
    max_gap = group['consecutive_missing'].max()

    # Interpolate the missing values for 'Daily.Mean.PM2.5.Concentration'
    group['Daily.Mean.PM2.5.Concentration'] = group['Daily.Mean.PM2.5.Concentration'].interpolate(method='linear')

    # After interpolation, assign the max gap to the entire group
    group['max_consecutive_missing'] = max_gap

    # Reset the index so 'Date' becomes a column again
    return group.reset_index().rename(columns={'index': 'Date'})

interpolated_data = valid_sites.groupby(['Site.ID', 'month', 'year']).apply(interpolate_monthly).reset_index(drop=True)

interpolated_data[interpolated_data["school_year"] == 2003]

# Drop the multi-index and reset index
interpolated_data.reset_index(drop=True, inplace=True)

interpolated_data.drop(columns=['count', 'month', 'day', 'aq_vuh'], inplace=True)
interpolated_data

interpolated_data.groupby(['Site.ID', 'school_year']).ngroups

interpolated_data.to_csv("./interpolated_aq_180.csv")
